+++ src/pages/index.jsx
@@ -1,4 +1,1 @@
-export { default as Login } from "../pages/login/Login";
-export { default as Register } from "../pages/register/Register";
-export { default as Restaurant } from "../pages/restaurant/Restaurant";
-export { default as RestaurantDetails } from "../pages/restaurantdetails/RestaurantDetails";
+export { default as DetailsRestaurants } from "../pages/detailrestaurants/DetailsRestaurants";
