+++ src/redux/actions/index.jsx
@@ -1,1 +1,1 @@
-export * from "./Restaurant";
+export * from "./DetailRestaurants";
